Switch to structs, change icon signature (#2)
This pull request includes several changes to the codebase. The main changes are:

Switching to using structs instead of classes for improved performance and memory efficiency.

Updating the icon signature to match the new struct implementation.

Switch to a valid protocol

This pull request updates the protocol used in the codebase from "mandadin-web" to "web+mandadin". This change ensures that the protocol follows the correct format and is compatible with other systems.

Add FsToolkit, add The Blunt Parser

This pull request adds the FsToolkit.ErrorHandling and TheBlunt packages to the project. These packages provide additional functionality and error handling capabilities.

Add Parser module for checkbox parsing

This pull request introduces a new Parser module for parsing checkboxes in the codebase. This module allows for more efficient and accurate parsing of checkbox data.

Add logging